# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedRise Medical Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vitals:
    get:
      operationId: listVitals
      tags: [vitals]
      summary: List vital signs
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of vitals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VitalSigns"
    post:
      operationId: createVitals
      tags: [vitals]
      summary: Record vital signs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VitalSignsInput"
      responses:
        "201":
          description: Vitals recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalSigns"
  /vitals/{id}:
    delete:
      operationId: deleteVitals
      tags: [vitals]
      summary: Delete a vitals record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── BILLING ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    VitalSigns:
      type: object
      required: [id, patientId, recordedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        consultationId:
          type: ["integer", "null"]
        bloodPressure:
          type: ["string", "null"]
        temperature:
          type: ["string", "null"]
        pulse:
          type: ["string", "null"]
        weight:
          type: ["string", "null"]
        height:
          type: ["string", "null"]
        oxygenSaturation:
          type: ["string", "null"]
        respiratoryRate:
          type: ["string", "null"]
        recordedBy:
          type: ["integer", "null"]
        recordedAt:
          type: string
    VitalSignsInput:
      type: object
      required: [patientId]
      properties:
        patientId:
          type: integer
        consultationId:
          type: integer
        bloodPressure:
          type: string
        temperature:
          type: string
        pulse:
          type: string
        weight:
          type: string
        height:
          type: string
        oxygenSaturation:
          type: string
        respiratoryRate:
          type: string
        recordedBy:
          type: integer

    # ── BILLING SCHEMAS ───────────────────────────────────────────────────
...